// src/app/models/product.model.ts
export interface Product {
  landingPageUrl: string;
  loyaltyPointsEnabled: boolean;
  adId: string;
  isPLA: boolean;
  barcode: string;
  product: string;
  name: string;
  rating: number;
  reviews: number;
  isFastFashion: boolean;
  futureDiscountedPrice: number;
  futureDiscountStartDate: string;
  discount: number;
  brand: string;
  searchImage: string;
  effectiveDiscountPercentageAfterTax: number;
  effectiveDiscountAmountAfterTax: number;
  buyButtonWinnerSkuId: number;
  buyButtonWinnerSellerPartnerId: number;
  relatedStylesCount: number;
  relatedStylesType: string;
  productVideos: any[];
  inventoryInfo: any[];
  sizes: string;
  images: any[];
  gender: string;
  color: string;
  discountLabel: string;
  discountDisplayLabel: string;
  additionalInfo: string;
  category: string;
  mrp: number;
  price: number;
  advanceOrderTag: string;
  colorVariantAvailable: boolean;
  productimagetag: string;
  listViews: number;
  discountType: string;
  tdBxGyText: string;
  catalogDate: string;
  season: string;
  year: string;
  isPersonalised: boolean;
  eorsPicksTag: string;
  personalizedCoupon: string;
  personalizedCouponValue: number;
  productMeta: string;
  systemAttributes: any[];
  attributeTagsPriorityList: any[];
  preferredDeliveryTag: string;
  deliveryPromise: string;
  expressTags: any[];
  couponData: any;
  productMetaData: any;
  mediaData: any[];
}